services/users-management/func_try_users/requirements.txt
azure-functions
azure-cli
python-dotenv
requests
msal